requirements.txt
fastapi==0.117.1
uvicorn==0.27.1
python-multipart==0.0.9
httpx==0.27.2
aiohttp==3.9.5
tenacity==8.2.2
orjson==3.9.9
python-dotenv==1.0.0
structlog==23.3.0
redis==5.1.1
pytest==7.4.0
pytest-asyncio==0.21.0
black==25.9.0
ruff==0.6.9
isort==5.13.2
flake8==7.1.1
mypy==1.11.2
pre-commit==3.8.0
coverage==7.6.1
pyyaml==6.0.2
rank-bm25==0.2.2
